requirements.txt
fastapi
uvicorn
jinja2
playwright
google-generativeai
python-dotenv
python-multipart
httpx
